Incorporation of new custom function addItemElement(). Adjusted images and changed display style
Incorporation of a brand new function called addItemElement which has the purpose of adding new divs with already existing styles  inside an existing div. It's intended to work as a receipt list of what items have been added to the shopping cart and how many of them are there. Some display changes have been made in order to display the items and it's contents optimally. Adjusted the names of the icons